etc/Code_Upload/GUI/pyproject.toml
[tool.poetry]
name = "MainGUI"
version = "0.1.0"
description = ""
authors = ["MiscDev <miscdev@mailworks.org>"]

[tool.poetry.dependencies]
python = "^3.7"
fuzzywuzzy = "^0.17.0"
python-Levenshtein = "^0.12.0"
pyside2 = "^5.12"
attrs = "^18.2"
loguru = "^0.2.5"
PyPDF2 = "^1.26"
reportlab = "^3.5"
joblib = "^0.13.2"
scikit-learn = "^0.20.2"
matplotlib = "^3.0"
requests = "^2.21"
execnet = "^1.6"
pyzmq = "^18.0"
tensorflow = "^1.13"
SigMF = "^0.0.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
nose = "^1.3"
autopep8 = "^1.4"
pylint = "^2.2"
sphinx = "^2.0"
wheel = "^0.33.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
